Native queries now return `Long` instead of `BigInteger`

+++ commons/src/test/java/org/hyades/persistence/UserRepositoryTest.java
@@ -2,11 +2,10 @@ package org.hyades.persistence;
 
 import io.quarkus.test.TestTransaction;
 import io.quarkus.test.junit.QuarkusTest;
-import org.junit.jupiter.api.Test;
-
 import jakarta.inject.Inject;
 import jakarta.persistence.EntityManager;
-import java.math.BigInteger;
+import org.junit.jupiter.api.Test;
+
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -24,16 +23,16 @@ class UserRepositoryTest {
     @TestTransaction
     @SuppressWarnings("unchecked")
     void testFindEmailsByTeam() {
-        final var teamIds = (List<BigInteger>) entityManager.createNativeQuery("""
+        final var teamIds = (List<Long>) entityManager.createNativeQuery("""
                 INSERT INTO "TEAM" ("NAME", "UUID") VALUES 
                     ('foo', 'ba38e779-e252-4033-8e76-156dc46cc7a6'),
                     ('bar', '507d8f3c-431d-47aa-929e-7647746d07a9')
                 RETURNING "ID";
                 """).getResultList();
-        final BigInteger teamFooId = teamIds.get(0);
-        final BigInteger teamBarId = teamIds.get(1);
+        final Long teamFooId = teamIds.get(0);
+        final Long teamBarId = teamIds.get(1);
 
-        final var managedUserIds = (List<BigInteger>) entityManager.createNativeQuery("""
+        final var managedUserIds = (List<Long>) entityManager.createNativeQuery("""
                 INSERT INTO "MANAGEDUSER" ("EMAIL", "PASSWORD", "FORCE_PASSWORD_CHANGE", "LAST_PASSWORD_CHANGE", "NON_EXPIRY_PASSWORD", "SUSPENDED") VALUES 
                     ('foo@managed.example.com', 'foo', false, NOW(), true, false),
                     ('bar@managed.example.com', 'bar', false, NOW(), true, false),
@@ -42,10 +41,10 @@ class UserRepositoryTest {
                     ('quux@example.com', 'quux', false, NOW(), true, false)
                 RETURNING "ID";
                 """).getResultList();
-        final BigInteger managedUserFooId = managedUserIds.get(0);
-        final BigInteger managedUserBarId = managedUserIds.get(1);
-        final BigInteger managedUserQuxId = managedUserIds.get(3);
-        final BigInteger managedUserQuuxId = managedUserIds.get(4);
+        final Long managedUserFooId = managedUserIds.get(0);
+        final Long managedUserBarId = managedUserIds.get(1);
+        final Long managedUserQuxId = managedUserIds.get(3);
+        final Long managedUserQuuxId = managedUserIds.get(4);
 
         entityManager.createNativeQuery("""
                         INSERT INTO "MANAGEDUSERS_TEAMS" ("MANAGEDUSER_ID", "TEAM_ID") VALUES
@@ -63,7 +62,7 @@ class UserRepositoryTest {
                 .setParameter("teamBarId", teamBarId)
                 .executeUpdate();
 
-        final var ldapUserIds = (List<BigInteger>) entityManager.createNativeQuery("""
+        final var ldapUserIds = (List<Long>) entityManager.createNativeQuery("""
                 INSERT INTO "LDAPUSER" ("EMAIL", "DN") VALUES 
                     ('foo@ldap.example.com', 'foo'),
                     ('bar@ldap.example.com', 'bar'),
@@ -72,10 +71,10 @@ class UserRepositoryTest {
                     ('quux@example.com', 'quux')
                 RETURNING "ID";
                 """).getResultList();
-        final BigInteger ldapUserFooId = ldapUserIds.get(0);
-        final BigInteger ldapUserBarId = ldapUserIds.get(1);
-        final BigInteger ldapUserQuxId = ldapUserIds.get(3);
-        final BigInteger ldapUserQuuxId = ldapUserIds.get(4);
+        final Long ldapUserFooId = ldapUserIds.get(0);
+        final Long ldapUserBarId = ldapUserIds.get(1);
+        final Long ldapUserQuxId = ldapUserIds.get(3);
+        final Long ldapUserQuuxId = ldapUserIds.get(4);
 
         entityManager.createNativeQuery("""
                         INSERT INTO "LDAPUSERS_TEAMS" ("LDAPUSER_ID", "TEAM_ID") VALUES
@@ -93,7 +92,7 @@ class UserRepositoryTest {
                 .setParameter("teamBarId", teamBarId)
                 .executeUpdate();
 
-        final var oidcUserIds = (List<BigInteger>) entityManager.createNativeQuery("""
+        final var oidcUserIds = (List<Long>) entityManager.createNativeQuery("""
                 INSERT INTO "OIDCUSER" ("EMAIL", "USERNAME") VALUES 
                     ('foo@oidc.example.com', 'foo'),
                     ('bar@oidc.example.com', 'bar'),
@@ -102,10 +101,10 @@ class UserRepositoryTest {
                     ('quux@example.com', 'quux')
                 RETURNING "ID";
                 """).getResultList();
-        final BigInteger oidcUserFooId = oidcUserIds.get(0);
-        final BigInteger oidcUserBarId = oidcUserIds.get(1);
-        final BigInteger oidcUserQuxId = oidcUserIds.get(3);
-        final BigInteger oidcUserQuuxId = oidcUserIds.get(4);
+        final Long oidcUserFooId = oidcUserIds.get(0);
+        final Long oidcUserBarId = oidcUserIds.get(1);
+        final Long oidcUserQuxId = oidcUserIds.get(3);
+        final Long oidcUserQuuxId = oidcUserIds.get(4);
 
         entityManager.createNativeQuery("""
                         INSERT INTO "OIDCUSERS_TEAMS" ("OIDCUSERS_ID", "TEAM_ID") VALUES
@@ -123,7 +122,7 @@ class UserRepositoryTest {
                 .setParameter("teamBarId", teamBarId)
                 .executeUpdate();
 
-        assertThat(repository.findEmailsByTeam(teamFooId.longValue())).containsExactlyInAnyOrder(
+        assertThat(repository.findEmailsByTeam(teamFooId)).containsExactlyInAnyOrder(
                 "foo@managed.example.com",
                 "bar@managed.example.com",
                 "foo@ldap.example.com",
@@ -132,7 +131,7 @@ class UserRepositoryTest {
                 "bar@oidc.example.com"
         );
 
-        assertThat(repository.findEmailsByTeam(teamBarId.longValue())).containsExactlyInAnyOrder(
+        assertThat(repository.findEmailsByTeam(teamBarId)).containsExactlyInAnyOrder(
                 "bar@managed.example.com",
                 "bar@ldap.example.com",
                 "bar@oidc.example.com",
